# Preencode attribute keys into utf8 bytes.