--- a/api/all/src/main/java/io/opentelemetry/api/internal/InternalAttributeKeyImpl.java
+++ b/api/all/src/main/java/io/opentelemetry/api/internal/InternalAttributeKeyImpl.java
@@ -3,18 +3,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.common;
+package io.opentelemetry.api.internal;
 
+import io.opentelemetry.api.common.AttributeKey;
+import io.opentelemetry.api.common.AttributeType;
+import java.nio.charset.StandardCharsets;
 import javax.annotation.Nullable;
 
 @SuppressWarnings("rawtypes")
-final class AttributeKeyImpl<T> implements AttributeKey<T> {
+public final class InternalAttributeKeyImpl<T> implements AttributeKey<T> {
 
   private final AttributeType type;
   private final String key;
+  private final byte[] keyUtf8;
   private final int hashCode;
 
-  private AttributeKeyImpl(AttributeType type, String key) {
+  private InternalAttributeKeyImpl(AttributeType type, String key) {
     if (type == null) {
       throw new NullPointerException("Null type");
     }
@@ -23,6 +27,7 @@ final class AttributeKeyImpl<T> implements AttributeKey<T> {
       throw new NullPointerException("Null key");
     }
     this.key = key;
+    this.keyUtf8 = key.getBytes(StandardCharsets.UTF_8);
     this.hashCode = buildHashCode(type, key);
   }
 
@@ -38,8 +43,8 @@ final class AttributeKeyImpl<T> implements AttributeKey<T> {
   // the class loader automatically resolves its super classes (interfaces), which in this case is
   // Context, which would be the same class (interface) being instrumented at that time,
   // which would lead to the JVM throwing a LinkageError "attempted duplicate interface definition"
-  static <T> AttributeKey<T> create(@Nullable String key, AttributeType type) {
-    return new AttributeKeyImpl<>(type, key != null ? key : "");
+  public static <T> AttributeKey<T> create(@Nullable String key, AttributeType type) {
+    return new InternalAttributeKeyImpl<>(type, key != null ? key : "");
   }
 
   @Override
@@ -52,13 +57,17 @@ final class AttributeKeyImpl<T> implements AttributeKey<T> {
     return key;
   }
 
+  public byte[] getKeyUtf8() {
+    return keyUtf8;
+  }
+
   @Override
   public boolean equals(@Nullable Object o) {
     if (o == this) {
       return true;
     }
-    if (o instanceof AttributeKeyImpl) {
-      AttributeKeyImpl<?> that = (AttributeKeyImpl<?>) o;
+    if (o instanceof InternalAttributeKeyImpl) {
+      InternalAttributeKeyImpl<?> that = (InternalAttributeKeyImpl<?>) o;
       return this.type.equals(that.getType()) && this.key.equals(that.getKey());
     }
     return false;