--- a/api/all/src/main/java/io/opentelemetry/api/internal/InternalAttributeKeyImpl.java
+++ b/api/all/src/main/java/io/opentelemetry/api/internal/InternalAttributeKeyImpl.java
@@ -10,6 +10,7 @@ import io.opentelemetry.api.common.AttributeType;
 import java.nio.charset.StandardCharsets;
 import javax.annotation.Nullable;
 
+/** Default AttributeKey implementation which preencodes to UTF8 for OTLP export. */
 @SuppressWarnings("rawtypes")
 public final class InternalAttributeKeyImpl<T> implements AttributeKey<T> {