--- a/api/all/src/test/java/io/opentelemetry/api/common/AttributeKeyTest.java
+++ b/api/all/src/test/java/io/opentelemetry/api/common/AttributeKeyTest.java
@@ -16,6 +16,7 @@ class AttributeKeyTest {
   @Test
   void equalsVerifier() {
     EqualsVerifier.forClass(InternalAttributeKeyImpl.class)
+        .withIgnoredFields("keyUtf8")
         .withCachedHashCode(
             "hashCode",
             "buildHashCode",