@Oberon00 2021-09-03
With this change, the `(String, ValueType)` overloads of the attribute setters (i.e. those that take a String instead of an AttributeKey, such as setAttribute on Span and SpanBuilder) might be worth deprecating, since there is now a significant performance overhead with these.