@jkwatson 2021-09-03
> With this change, the `(String, ValueType)` overloads of the attribute setters (i.e. those that take a String instead of an AttributeKey, such as setAttribute on Span and SpanBuilder) might be worth deprecating, since there is now a significant performance overhead with these.

We've already got some pretty strong warnings on these methods about not using them if you don't have to. I don't think we should deprecate them, because they definitely are convenient for people who don't care about hyper performance optimization.